+++ src/heaps/triheap.h
@@ -48,7 +48,7 @@ class TriHeapNode {
         unsigned int extra;    
         unsigned int dim;
 
-        float key;
+        double key;
         unsigned int item;
 };
 
@@ -79,8 +79,8 @@ class TriHeap : public Heap {
         ~TriHeap();
 
         unsigned int deleteMin();
-        void insert(unsigned int item, float k);
-        void decreaseKey(unsigned int item, float newValue);
+        void insert(unsigned int item, double k);
+        void decreaseKey(unsigned int item, double newValue);
         unsigned int nItems() const { return itemCount; }
 
         long int nComps() const { return compCount; }
